app/requirements.txt
flask
paho-mqtt
prometheus_client
requests
psutil
docker